2012-03-16 10:02 Refine declaration of evaluation strategy
OK, here's the high-level overview of how ACE is supposed to work.
NRS works slightly differently, which we can ignore for now since
NRS and API are incompatible.

gsvalues are either unboxed values (high bit *on*) or pointers
(high bit *off*).  Unboxed values are NFs by definition, so we can
ignore them for now.

Every pointer points into a /megablock/, which is a 256KW block of memory.
(1KW = 1024 pointers, so on a 32-bit machine 256KW = 1MB).  The first word
of a megablock is a /registered block type/, which is a pointer to a
(usually) statically-allocated block of memory containing a C struct.
The first member of that C struct is the "evaluator" for pointers into the block.
(Remember: the pointer is a "Global Script value", not the pointed-at object).
The "evaluator" is supposed to return the state of the value; basically:

 * Not yet being evaluated (gstythunk)
 * In the process of being evaluated (gstystack)
 * Already evaluated (many of the others)

NB: We should really add a "blocked" status there.
It would be good to return more information, but that's a secondary step.

For pointers into the Global Script heap (i.e., actual redexes), the
"evaluator" is supposed to arrange for evaluation of the expression,
/in parallel with whatever the caller's thread is doing/.  To handle this,
ACE has its own pool of OS threads, and all calls to the evaluator send
messages to that pool.  For redexes hanging out in the heap, ACE is supposed
to create a new evaluation stack (new "thread" or /eval/) with the update
frame for the redex at the bottom of the stack.  (Unboxed values need to be
handled specially --- I have no idea how that should work.  Values *have*
to be boxed somehow to be handed back to the client, yes?)  Then the pool
of threads picks up that eval among the others for evaluation.

Once a redex has been moved onto an evaluation stack, we replace the redex
with a pointer to the update frame on the evaluation stack (this is what
GHC calls "black holing", although I don't think that's the right
conceptual framework for what ACE is doing).  Calling the evaluator again, on
a redex with a pointer to an evaluation stack, does nothing (maybe it
does a one-time bump in priority for that eval) and returns gstystack.

Eventually, ACE will (hopefully) finish reducing the stack to WHNF, at which
point it will update the original value in the heap with an indirection to
the actual value.  After this point, the evaluator will return "gstyindir",
and after the next GC it will return "gstyconstr" or something.

Callers to the evaluator should follow this code pattern:

 * Fetch the evaluator
 * Short-circuit calls to "our own" evaluator
 * Otherwise, call the evaluator

For example, when ACE itself goes to force a thunk, "its own" evaluator
is the standard ACE evaluator, which could maybe be used directly but it's
faster to just push the new stack directly atop the current one.  (And it
confuses ACE's scheduler less to do it that way).  Similar short-circuiting
makes sense in other cases.  Not sure the best way to handle that.