[fv-names] Check register names supplied for free variables
Check that the register name of supplied values for free variables is the same as the name of those variables in the sub-code item.
Fix code in the distribution that violates that rule.

This will enable us to use the names in the sub-code item instead of the supplied names further on.